simple-agent/pyproject.toml
[project]
name = "simple-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-groq>=1.1.2",
    "langchain-mcp-adapters>=0.2.2",
    "langgraph>=1.1.6",
    "python-dotenv>=1.2.2",
]
